+++ Ya.hs
@@ -1,3 +1,4 @@
 module Ya (module Exports) where
 
 import Ya.Algebra as Exports
+import Ya.Program as Exports